Fixed JsonAny handling, added more webview tests

+++ packages/vscode-messenger-webview/src/messenger.ts
@@ -4,7 +4,7 @@
  * terms of the MIT License, which is available in the project root.
  ******************************************************************************/
 
-import { createMessage, isRequestMessage, isResponseMessage, MessageParticipant, MessengerAPI, NotificationHandler, NotificationType, RequestHandler, RequestType } from 'vscode-messenger-common';
+import { createMessage, isRequestMessage, isResponseMessage, JsonAny, MessageParticipant, MessengerAPI, NotificationHandler, NotificationType, RequestHandler, RequestType } from 'vscode-messenger-common';
 import { acquireVsCodeApi, VsCodeApi } from './vscode-api';
 
 export class Messenger implements MessengerAPI {
@@ -18,7 +18,7 @@ export class Messenger implements MessengerAPI {
         this.vscode = vscode ?? acquireVsCodeApi();
     }
 
-    sendRequest<P, R>(type: RequestType<P, R>, receiver: MessageParticipant, params: P): Promise<R> {
+    sendRequest<P extends JsonAny, R>(type: RequestType<P, R>, receiver: MessageParticipant, params: P): Promise<R> {
         const msgId = this.createMsgId();
         const result = new Promise<R>((resolve, reject) => {
             this.requests.set(msgId, { resolve, reject });
@@ -28,53 +28,53 @@ export class Messenger implements MessengerAPI {
         return result;
     }
 
-    sendNotification<P>(type: NotificationType<P>, receiver: MessageParticipant, params: P): void {
+    sendNotification<P extends JsonAny>(type: NotificationType<P>, receiver: MessageParticipant, params: P): void {
         this.vscode.postMessage(createMessage(this.createMsgId(), type, receiver, params));
     }
 
-    onRequest<P, R>(type: RequestType<P, R>, handler: RequestHandler<P, R>): void {
+    onRequest<P extends JsonAny, R>(type: RequestType<P, R>, handler: RequestHandler<P, R>): void {
         this.handlerRegistry.set(type.method, handler);
     }
 
-    onNotification<P>(type: NotificationType<P>, handler: NotificationHandler<P>): void {
+    onNotification<P extends JsonAny>(type: NotificationType<P>, handler: NotificationHandler<P>): void {
         this.handlerRegistry.set(type.method, handler);
     }
 
-    start(): void {
+    start(): Messenger {
         window.addEventListener('message', event => {
             const data = event.data;
-            if(isResponseMessage(data)) {
-                console.log(`Response message: ${data.id} `);
+            if (isResponseMessage(data)) {
+                console.debug(`Response message: ${data.id} `);
                 const request = this.requests.get(data.id);
-                if(data.error) {
-                    request.reject(data.error);
-                } else {
-                    request.resolve(data.result);
-                }
-                this.requests.delete(data.id);
-            } else if(isRequestMessage(data)) {
-                console.log(`Request message: ${data.id} `);
-                const handler = this.handlerRegistry.get(data.method);
-                if(handler) {
-                    const parameter = data.params;
-                    // TODO check this and handle other cases
-                    if (typeof parameter === 'string' || parameter instanceof String) {
-                        handler(JSON.parse(data.params as string));
+                if (request) {
+                    if (data.error) {
+                        request.reject(data.error);
                     } else {
-                        handler(data.params);
+                        request.resolve(data.result);
                     }
+                    this.requests.delete(data.id);
+                } else {
+                    console.warn(`Received response for untracked message id: ${data.id}. Receiver was: ${data.receiver?.webviewId ?? data.receiver?.webviewType}`);
+                    return;
+                }
+            } else if (isRequestMessage(data)) {
+                console.debug(`Request message: ${data.id} `);
+                const handler = this.handlerRegistry.get(data.method);
+                if (handler) {
+                    handler(data.params);
                 }
             }
         });
+        return this;
     }
 
     private id = 0;
 
-    private createMsgId(): string {
+    protected createMsgId(): string {
         // messenger is created each time a view gets visible.
-        // TODO Use uuid library
-        const uuid = Date.now().toString(36) + Math.random().toString(36).substr(2);
-        return 'viewMsgId_' + this.id++ + '_' + uuid;
+        const cryptoRand = window.crypto.getRandomValues(new Uint8Array(10));
+        const rand = Array.from(cryptoRand).map(b => b.toString(16)).join('');
+        return 'viewMsgId_' + this.id++ + '_' + rand;
     }
 
 }
